Fix snooze button dismissing alarm
