+++ lib/alarm/widgets/notification_actions/buttons_notification_action.dart
@@ -15,7 +15,7 @@ class ButtonsNotificationAction extends StatelessWidget {
       mainAxisAlignment: MainAxisAlignment.center,
       children: [
         CardContainer(
-          onTap: onDismiss,
+          onTap: onSnooze,
           child: Padding(
             padding: const EdgeInsets.all(32.0),
             child:
